port/wasm/artifact-explorer: integrate eiafx-data
This commit only ports the app to use the data where possible. Effects and
recipes will be integrated in a separate commit

This commit also de-link the mission/artifact card headings -- they were links
to themselves which is pointless and makes it hard to copy the textual
content. Should be a separate commit, but don't want to further split hunks, oh
well.