lib/prophecy_eggs: use backup.game.eggsOfProphecy as total...
and derive total contract PEs from that.

Rationale: Sometimes when a player has started a half-finished legacy contract,
and they have completed a goal that wasn't a PE goal the last time but has
turned into a PE goal this time, the PE reward should be reimbursed, but the PE
count may not update immediately. It may update at some random point. In this
case, to best avoid confusion, we use the reported total
backup.game.eggsOfProphecy, and deduce the number of contract PEs from that.

See the following bug report:
https://discord.com/channels/869885242801029150/869945876313944074/914077000824659979
https://discord.com/channels/@me/849156772999462922/914078132309467196