ui: BaseSelectFilterable: switch to JSX for render function
Unfortunately, <template> doesn't work with the generically typed component
pattern; in development mode everything's fine, but in production mode, the
render function compiled from the template is simply missing; see
https://1f91b36168d7f0e48c0c697006b826515f4e5b6b--wasmegg.netlify.app/artifact-sandbox
for the catastrophic result (component with no render function and therefore
renders to <!---->).

As such we resort to defining the render function directly in defineComponent
with JSX. This is not ideal since (1) another plugin required
(@vitejs/plugin-vue-jsx); (2) mixed codebase; (3) volar/eslint is not exactly
happen with the JSX for some reason.

Note that the codesandbox example using @vue/cli-service does work even in
production. Why it doesn't work with vite is beyond me. I'll need to consult
with an expert. If the problem is resolved, revert to the
1f91b36168d7f0e48c0c697006b826515f4e5b6b version.